Refactor: enum 타입 폴더로 관리
